chore: Refactor Actions
With the new implementation for
the actions it is more easy to implement actions.

An action is now capable of updating the argument
parser and can be created dynamically
