+++ src/pdm_bump/actions/__init__.py
@@ -9,3 +9,11 @@
 # This file is published using the MIT license.
 # Refer to LICENSE for more information
 #
+
+from .base import actions
+from .increment import *  # noqa
+from .preview import *  # noqa
+
+__all__ = [
+    "actions",
+]
